feat: initial work on PWA geolocation

+++ Blazor/Dto/LocationDto.cs
@@ -2,7 +2,7 @@
 {
     public class LocationDto
     {
-        public long Longitude {  get; set; }    
-        public long Latitude {  get; set; } 
+        public double Longitude {  get; set; }    
+        public double Latitude {  get; set; } 
     }
 }
